Refactor admission checker unit tests
